Added "Nullable" and "Value Type" options to the data type
If the "Nullable" option is enabled, the underlying value converter will return null if the saved property value is null or white space.

The "Value Type" option allows setting the C# type returned by the value converter. Default is "IHtmlEncodedString", but the options allows chaching the value type to either "IHtmlContent" or "string".